refactor: move project text to external file
