+++ src/components/staffing/ProjectCell.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Cell } from 'fixed-data-table'
 import { IGNORED_PROJECT_NAMES, MAP_PROJECT_NAMES } from '../../constants'
+import ProjectCellText from './ProjectCellText'
 
 export default class ProjectCell extends React.Component {
   static propTypes = {
@@ -29,30 +30,20 @@ export default class ProjectCell extends React.Component {
     const upcomingProjects = projects.filter(
       (project) => { return !currentProjects.includes(project) }
     )
-    const renamedCurrentProjects = currentProjects.map(
-      (project) => {
-        return MAP_PROJECT_NAMES[project]
-        ? MAP_PROJECT_NAMES[project]
-        : project
-      }
-    )
-    const currentProjectsText = renamedCurrentProjects.join(', ')
-    const separator = currentProjects.length > 0 && upcomingProjects.length > 0 ? ', ' : ''
-    const upcomingProjectsText = upcomingProjects.join(', ')
     const isIndividualProject = data[rowIndex].projects === undefined
     const individualProjectText = isIndividualProject ? data[rowIndex][field] : ''
     const individualProjectRenamedText = MAP_PROJECT_NAMES[individualProjectText]
     ? MAP_PROJECT_NAMES[individualProjectText]
     : individualProjectText
-    const cellText = isIndividualProject ? individualProjectRenamedText : (
-      <div>
-        <b>
-          {currentProjectsText}
-        </b>
-        { separator }
-        { upcomingProjectsText }
-      </div>
+    const cellText = isIndividualProject
+    ? individualProjectRenamedText
+    : (
+      <ProjectCellText
+        currentProjects={currentProjects}
+        upcomingProjects={upcomingProjects}
+      />
     )
+
     return (
       <Cell
         {...props}
